requirements.txt
absl-py==0.12.0
alabaster==0.7.12
alembic==1.5.8
argon2-cffi
arrow==1.1.0
async-generator
attrs
Babel==2.9.1
backcall
beautifulsoup4==4.9.3
bibtexparser==1.2.0
bleach
brotlipy==0.7.0
cachetools==4.2.1
category-encoders==2.2.2
certifi==2020.12.5
cffi
chardet
click==7.1.2
click-plugins==1.1.1
cliff==3.7.0
cligj
cmaes==0.8.2
cmd2==1.5.0
colorama
colorlog==5.0.1
cryptography
cycler==0.10.0
decorator
defusedxml
distro==1.5.0
docopt==0.6.2
docutils==0.16
entrypoints==0.3
fake-useragent==0.1.11
Fiona==1.8.19
free-proxy==1.0.2
fsspec==0.8.7
future==0.18.2
gitdb==4.0.7
GitPython==3.1.14
google-auth==1.29.0
google-auth-oauthlib==0.4.4
greenlet==1.0.0
grpcio==1.37.0
gunicorn==20.1.0
idna
imagesize==1.2.0
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
jedi==0.17.0
Jinja2
joblib
jsonschema
kiwisolver
lightgbm==3.2.1
llvmlite==0.36.0
lxml==4.6.3
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib
mistune==0.8.4
munch==2.5.0
nest-asyncio
notebook
numba==0.53.1
numpy
oauthlib==3.1.0
olefile
optuna==2.7.0
packaging
pandas
pandocfilters
parso
patsy==0.5.1
pbr==5.5.1
pickleshare
Pillow
pipreqs==0.4.10
plotly
prettytable==2.1.0
prometheus-client
prompt-toolkit
protobuf==3.15.8
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser
Pygments
pyOpenSSL
pyparsing
pyperclip==1.8.2
pyproj
pyreadline3==3.3
pyrsistent
PySocks
python-dateutil
python-dotenv==0.17.1
python-editor==1.0.4
pytz
qtconsole
QtPy==1.9.0
requests
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7.2
scikit-learn
scikit-optimize==0.8.1
scipy
seaborn==0.11.1
selenium==3.141.0
Send2Trash
Shapely==1.7.1
six
sktime==0.6.0
smmap==4.0.0
snowballstemmer==2.1.0
soupsieve==2.2.1
Sphinx==4.0.2
sphinx-rtd-theme==0.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.9
statsmodels==0.12.2
stem==1.8.0
stevedore==3.3.0
terminado==0.9.4
testpath
threadpoolctl
torchmetrics==0.2.0
tornado
traitlets
typing-extensions
urllib3
wcwidth
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton
wincertstore==0.2
yarg==0.1.9
zipp
dash==1.20.0
dash-bootstrap-components==0.12.2
dash-core-components==1.16.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
geopandas==0.9.0
gunicorn==20.1.0
